% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{mergebamsR-package}
\alias{mergebamsR}
\alias{mergebamsR-package}
\title{mergebamsR: merging bams with R}
\description{
not that complicated
}
\seealso{
Useful links:
\itemize{
  \item \url{https://furlan-lab.github.io/mergebamsR/}
}

}
\author{
\strong{Maintainer}: Scott Furlan \email{scottfurlan@gmail.com} (\href{https://orcid.org/0000-0003-3007-9840}{ORCID})

}
