% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergebamsR.R
\name{subsetbam}
\alias{subsetbam}
\title{Subset a BAM file based on specific tags}
\usage{
subsetbam(
  inputbam,
  tags,
  outputbams,
  prefixes = NULL,
  TAG = "CB",
  cores = 1,
  verbose = F
)
}
\arguments{
\item{inputbam}{A character string specifying the path to the input BAM file.}

\item{tags}{A character vector of tags to filter the reads by.}

\item{outputbams}{A character vector specifying the paths to the output BAM files.}

\item{prefixes}{Optional; a character vector of prefixes to add to the read names in the output BAM files.
If NULL, no prefixes are added. Defaults to NULL.}

\item{TAG}{Optional; the tag used for subselection, default is "CB".}

\item{cores}{Optional; the number of cores to use for the process, default is 1.}

\item{verbose}{Option to increase verbosity of output}
}
\value{
No return value, called for side effects.
}
\description{
This function subselects reads from a BAM file based on the provided tags and
writes the subselected reads into separate output BAM files. The subselection
is done using a specified tag (default "CB"). This function can utilize multiple cores
to perform the operation more efficiently.
}
\details{
It's important that the length of \code{tags} is equal to the length of \code{outputbams}.
If \code{prefixes} is provided, its length must also match the length of \code{outputbams}.
If any of these conditions is not met, the function will stop and throw an error.
}
\examples{
subsetbam(inputbam = "path/to/input.bam", 
          tags = c("TAG1", "TAG2"), 
          outputbams = c("path/to/output1.bam", "path/to/output2.bam"),
          prefixes = c("prefix1", "prefix2"),
          TAG = "CB",
          cores = 2)

}
