% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{subsetbam_rust_helper}
\alias{subsetbam_rust_helper}
\title{subsetbam_rust}
\usage{
subsetbam_rust_helper(
  inputbam,
  features,
  outputbams,
  tag,
  cores,
  field,
  dump_bam
)
}
\description{
subsetbam_rust
}
\keyword{internal}
