% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergebamsR.R
\name{subsetbam}
\alias{subsetbam}
\title{Subset BAM Files Based on Features}
\usage{
subsetbam(
  inputbam,
  features,
  outputbams,
  field = c("tag", "name"),
  dump_bam = NA,
  TAG = "CB",
  cores = 1,
  verbose = F,
  split_bam = F
)
}
\arguments{
\item{inputbam}{A string specifying the path to the input BAM file.}

\item{features}{A list of features (e.g., cell barcodes) to subset.}

\item{outputbams}{A vector of strings specifying the paths to the output BAM files.}

\item{field}{A string specifying the field to use for subsetting. Must be either \code{"tag"} ( use with TAG 'cb' for cell barcode) or \code{"name"}. Default is \code{"tag"}.}

\item{dump_bam}{A string specifying the path to the dump BAM file. Default is \code{NA}.}

\item{TAG}{A string specifying the BAM tag to use for subsetting. Default is \code{"CB"}.}

\item{cores}{An integer specifying the number of cores to use for parallel processing. Default is \code{1}.}

\item{verbose}{A logical indicating whether to print detailed messages. Default is \code{FALSE}.}

\item{split_bam}{A logical indicating whether to split the BAM file across multiple cores. Default is \code{FALSE}.}
}
\value{
None
}
\description{
This function subsets a BAM file based on provided features (e.g., cell barcodes) and outputs one or more BAM files.
}
\details{
It's important that the length of \code{features} is equal to the length of \code{outputbams}.
}
